add mobile tabs for trends
