+++ src/components/Trends/Explore/TrendsExploreSearch.module.scss
@@ -8,7 +8,6 @@
 
   @include responsive('phone-xs', 'phone') {
     width: 100%;
-    margin-left: -16px;
   }
 }
