+++ src/components/Trends/TrendsTable/TrendsTable.js
@@ -1,4 +1,5 @@
 import React, { PureComponent } from 'react'
+import { compose } from 'redux'
 import Table from 'react-table'
 import cx from 'classnames'
 import { Link } from 'react-router-dom'
@@ -20,26 +21,33 @@ import InsightCardSmall from '../../../components/Insight/InsightCardSmall'
 import ExplanationTooltip from '../../../components/ExplanationTooltip/ExplanationTooltip'
 import ConditionalWrapper from './ConditionalWrapper'
 import styles from './TrendsTable.module.scss'
+import withSizes from 'react-sizes'
+import { mapSizesToProps } from '../../../utils/withSizes'
+import { isDesktop } from 'react-sizes/src/presets'
 
-const columns = [
+const MOBILE_COLUMNS = [
   {
-    Header: '#',
-    accessor: 'index',
-    width: 35,
-    headerClassName: styles.headerIndex
-  },
-  {
-    Header: 'Word',
+    Header: 'Trending words',
     accessor: 'word'
   },
   {
-    Header: 'Hype score',
+    Header: 'Trending score',
     accessor: 'score'
   },
   {
     Header: 'Social volume',
     accessor: 'volume'
   }
+]
+
+const DESKTOP_COLUMNS = [
+  {
+    Header: '#',
+    accessor: 'index',
+    width: 35,
+    headerClassName: styles.headerIndex
+  },
+  ...MOBILE_COLUMNS
 ]
 
 const NumberCircle = ({ className, ...props }) => (
@@ -212,7 +220,8 @@ class TrendsTable extends PureComponent {
       username,
       selectTrend,
       selectedTrends,
-      trendConnections
+      trendConnections,
+      isDesktop
     } = this.props
 
     const tableData = trendWords.map(({ word, score }, index) => {
@@ -260,6 +269,8 @@ class TrendsTable extends PureComponent {
       }
     })
 
+    const baseColumns = isDesktop ? DESKTOP_COLUMNS : MOBILE_COLUMNS
+
     return (
       <PanelWithHeader
         header={header}
@@ -274,10 +285,10 @@ class TrendsTable extends PureComponent {
           data={tableData}
           columns={
             small
-              ? columns.slice(0, 2)
+              ? baseColumns.slice(0, 2)
               : hasActions
-                ? columns.concat(this.getActionButtons())
-                : columns
+                ? baseColumns.concat(this.getActionButtons())
+                : baseColumns
           }
           showPagination={false}
           defaultPageSize={10}
@@ -300,4 +311,7 @@ const mapStateToProps = ({
   username
 })
 
-export default connect(mapStateToProps)(TrendsTable)
+export default compose(
+  connect(mapStateToProps),
+  withSizes(mapSizesToProps)
+)(TrendsTable)
