+++ src/pages/Trends/LabsTrendsPage.module.scss
@@ -2,6 +2,10 @@
 
 .insights {
   margin-top: 40px;
+
+  @include responsive('phone', 'phone-xs') {
+    padding: 0 16px;
+  }
 }
 
 .trendsWrapper {
@@ -9,7 +13,7 @@
   margin-bottom: 2rem;
 
   @include responsive('phone', 'phone-xs') {
-    margin: 0 16px;
+    padding: 0 16px;
   }
 }
 
@@ -17,6 +21,10 @@
   display: flex;
   align-items: baseline;
   justify-content: center;
+
+  @include responsive('phone', 'phone-xs') {
+    margin: 0 auto 0 0;
+  }
 }
 
 .title {
@@ -28,6 +36,8 @@
 
   @include responsive('phone', 'phone-xs') {
     @include text('h4');
+
+    margin: 0;
   }
 }
 
@@ -41,6 +51,15 @@
 
   &__form {
     width: 430px;
+
+    @include responsive('phone', 'phone-xs') {
+      width: 100%;
+    }
+  }
+
+  @include responsive('phone', 'phone-xs') {
+    padding: 0 16px;
+    margin-bottom: 16px;
   }
 }
 
@@ -57,6 +76,10 @@
 
 .header {
   display: flex;
+
+  @include responsive('phone', 'phone-xs') {
+    padding: 0 16px;
+  }
 }
 
 .back {
